AAAAAAAAAA

+++ main.py
@@ -5,7 +5,7 @@ import os
 from sys import argv
 
 # if env file doesn't exist or argv[1] == "setup", create it
-if not os.path.isfile("env.json") or argv[1] == "setup":
+if not os.path.isfile("env.json") or (len(argv) > 1 and argv[1] == "setup"):
     print("Creating config file...")
     with open("env.json", "w") as f:
         f.write(json.dumps({"DISCORD_TOKEN": "bot token here" if argv[2] == None else argv[2], "ADDITIONAL_OWN_PERMS": []}, indent=4))
